refactor: #240 유효성 검사 확인 모달 컴포넌트 분리
