+++ src/components/index.ts
@@ -149,8 +149,9 @@ import Modal from './modal/Modal';
 import WarnRoleDelete from './recruit/recruitDetail/modal/warnRoleDelete/WarnRoleDelete';
 import Footer from './footer/Footer';
 import PortfolioModal from './portfolio/modal/PortfolioModal';
-import WarnRoleCount from './recruit/create/modal/WarnRoleCount';
+import WarnRoleCount from './recruit/create/modal/warnRoleCount/WarnRoleCount';
 import AccountDelete from './recruit/recruitDetail/modal/accountDelete/AccountDelete';
+import FinalValidationModal from './recruit/create/modal/finalValidationModal.tsx/FinalValidationModal';
 
 export {
 	Header,
@@ -297,4 +298,5 @@ export {
 	PortfolioModal,
 	WarnRoleCount,
 	AccountDelete,
+	FinalValidationModal,
 };
